bitewise_backend/requirements.txt
Flask
Flask-Cors
flask-sqlalchemy
python-dotenv
pytesseract
Pillow
numpy
requests
google-genai
deep-translator
opencv-python
pytest
